[bugfix] crash on navigation
