+++ app/src/main/java/ir/composenews/ui/MainViewModel.kt
@@ -3,9 +3,9 @@ package ir.composenews.ui
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import ir.composenews.base.BaseViewModel
-import ir.composenews.base.MainContract
 import ir.composenews.core_test.dispatcher.DispatcherProvider
-import ir.composenews.domain.model.Market
+import ir.composenews.navigation.MainContract
+import ir.composenews.uimarket.model.MarketModel
 import ir.composenews.utils.ContentType
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -22,11 +22,11 @@ class MainViewModel @Inject constructor(
     override val state: StateFlow<MainContract.State> = mutableState.asStateFlow()
     override fun event(event: MainContract.Event) {
         when (event) {
-            is MainContract.Event.SetMarket -> setMarket(event.market as Market?, event.contentType)
+            is MainContract.Event.SetMarket -> setMarket(event.market, event.contentType)
         }
     }
 
-    private fun setMarket(market: Market?, contentType: ContentType) = viewModelScope.launch {
+    private fun setMarket(market: MarketModel?, contentType: ContentType) = viewModelScope.launch {
         mutableState.emit(
             mutableState.value.copy(
                 market,
